Fix some "warning: dereferencing type-punned pointer will break strict-aliasing rules"
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@1371 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/luatexdir/ptexlib.h
@@ -203,7 +203,7 @@ extern integer write_tounicode(char **, char *);
 /* utils.c */
 extern boolean str_eq_cstr(strnumber, char *);
 extern char *makecstring(integer);
-extern char *makeclstring(integer, int *);
+extern char *makeclstring(integer, size_t *);
 extern void print_string(char *j);
 extern void append_string(char *s);
 extern void getcreationdate(void);
